refactor(numeral-date): move out of external folder
BREAKING CHANGE: NumeralDate has moved to src/components/numeral-date
