+++ src/components/numeral-date/index.d.ts
