+++ src/components/numeral-date/numeral-date.d.ts
