+++ src/components/numeral-date/numeral-date.stories.mdx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
-import StyledSystemProps from "../../../../.storybook/utils/styled-system-props";
+import StyledSystemProps from "../../../.storybook/utils/styled-system-props";
 import NumeralDate from ".";
 
 <Meta
@@ -21,7 +21,7 @@ For dates close to today, we advise the use of a standard datepicker. If you req
 Import `NumeralDate` into the project.
 
 ```javascript
-import NumeralDate from "carbon-react/lib/__experimental__/components/numeral-date";
+import NumeralDate from "carbon-react/lib/components/numeral-date";
 ```
 
 ## Examples
