+++ src/components/numeral-date/numeral-date.style.js
@@ -1,10 +1,10 @@
 import styled, { css } from "styled-components";
-import StyledValidationIcon from "../../../components/validations/validation-icon.style";
-import StyledIconSpan from "../../../__internal__/input-icon-toggle/input-icon-toggle.style";
-import { baseTheme } from "../../../style/themes";
-import StyledFormField from "../form-field/form-field.style";
-import StyledInputPresentantion from "../input/input-presentation.style";
-import StyledIcon from "../../../components/icon/icon.style";
+import StyledValidationIcon from "../validations/validation-icon.style";
+import StyledIconSpan from "../../__internal__/input-icon-toggle/input-icon-toggle.style";
+import { baseTheme } from "../../style/themes";
+import StyledFormField from "../../__experimental__/components/form-field/form-field.style";
+import StyledInputPresentantion from "../../__experimental__/components/input/input-presentation.style";
+import StyledIcon from "../icon/icon.style";
 
 export const StyledNumeralDate = styled.div`
   display: inline-flex;
